Revert `full_text_field` to `ocr_text_tesi`
This commit will revert the `full_text_field` to `ocr_text_tesi` in the
`CatalogController`.  We were seeing a behavior with the current IIIF
Print gem where the `full_text_field` was not behaving as expected when
searching for multiple terms.  Seems like it was because the
`all_text_tsimv` field only had the words while Essi's `ocr_text_tesi`
was the full Alto XML.